Add recursive list reversal project.
Meant to be done in lisp-like languages, this is a student exercise with the goal of using recursion (and as much as possible, tail recursion) to reverse a list of items.  Any item in the list may be another list that should also be reversed, etc.

Done in F#, which is missing that type of list, so I had to create that as well.  Basic tree model idea to replace nested lists was found on the internet (`type 'a tree`), rest was done by myself.